Move masking tests out of lib code
